+++ test/graph/GraphQLTest.ts
@@ -147,9 +147,14 @@ query Repos($teamId: ID!, $offset: Int!) {
         assert.equal(query, "subscription Test { Repo { name owner org { team { name } } channels { name } }}");
     });
 
-    it("should successfully inline fragments from graphql folder", () => {
+    it("should successfully inline fragments from graphql folder (file name lookup)", () => {
         const query = GraphQL.subscription("subscriptionWithFragmentInGraphql");
         assert.equal(query, "subscription Test { Repo { name owner org { team { name } } channels { name } }}");
+    });
+
+    it("should successfully inline fragments from graphql folder (operation lookup)", () => {
+        const query = GraphQL.subscription("TestFooBar");
+        assert.equal(query, "subscription TestFooBar { Repo { name owner org { team { name } } channels { name } }}");
     });
 
     it("should successfully inline fragments with path and fragmentDir", () => {